moving Frame to own class, moving grouper from canmatrix to utils, refactorying tests according to new strucutre.
