+++ src/canmatrix/canmatrix.py
@@ -79,12 +79,7 @@ defaultFloatFactory = decimal.Decimal  # type: typing.Callable[[typing.Any], can
 
 
 
-# https://docs.python.org/3/library/itertools.html
-def grouper(iterable, n, fillvalue=None):
-    """Collect data into fixed-length chunks or blocks."""
-    # grouper('ABCDEFG', 3, 'x') --> ABC DEF Gxx"
-    args = [iter(iterable)] * n
-    return zip_longest(*args, fillvalue=fillvalue)
+
 
 
 def unpack_bitstring(length, is_float, is_signed, bits):
@@ -146,11 +141,7 @@ def pack_bitstring(length, is_float, value, signed):
 
 
 
-@attr.s(eq=False)
-class AutosarE2EProperties(object):
-    profile = attr.ib(default=None)  # type: str
-    data_ids = attr.ib(default=None) # type: List[int]
-    data_length = attr.ib(default=None)  # type: int
+
 
 
 @attr.s(eq=False)
